Ryhmän luominen syöttää nyt myös groupMember tauluun ryhmän id:n, käyttäjän id:n, sekä ryhmän luojalle rooliksi admin
